Fixes Typescript errors
