+++ src/app/users/onboarding/page.spec.tsx
@@ -3,12 +3,13 @@ import {fireEvent, render, screen, waitFor} from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import {beforeEach, describe, expect, it, vi} from "vitest";
 
-import OnboardingPage, {handleStep1, handleStep2, InitialForm, RequiredCheckbox, SequentialForm} from "./page";
+import OnboardingPage, {handleStep1, handleStep2, InitialForm, SequentialForm} from "./page";
 
 import * as CreateModule from "@/app/users/actions/create";
 import * as Clerk from "@clerk/nextjs";
 import {UserResource, UseSignUpReturn, UseUserReturn} from "@clerk/types";
 import {CreateUserResponse} from "@/app/users/actions/create";
+import {RequiredCheckbox} from "@/components/requiredCheckbox";
 
 vi.mock("@/app/users/actions/create");
 
@@ -203,9 +204,9 @@ describe("Onboarding components and functions", () => {
         });
 
         it("calls onError and returns false when updateUser fails", async () => {
-            updateUserMock.mockResolvedValueOnce({ ok: false, errors: ["nope"] } as any);
+            updateUserMock.mockResolvedValueOnce({ ok: false, errors: ["nope"] } as unknown as CreateUserResponse);
             const mockedReload = vi.fn();
-            const user = { id: "u1", reload: mockedReload } as any;
+            const user = { id: "u1", reload: mockedReload } as unknown as UserResource;
             const onError = vi.fn();
             const fd = new FormData();
             const result = await handleStep2(fd, onError, user);
@@ -216,12 +217,11 @@ describe("Onboarding components and functions", () => {
         it("handles thrown errors by calling onError and returning false", async () => {
             updateUserMock.mockImplementationOnce(() => { throw new Error("ohno"); });
             const mockedReload = vi.fn();
-            const user = { id: "u1", reload: mockedReload } as any;
+            const user = { id: "u1", reload: mockedReload } as unknown as UserResource;
             const onError = vi.fn();
             const fd = new FormData();
             const result = await handleStep2(fd, onError, user);
             expect(result).toBe(false);
-            expect(onError).toHaveBeenCalledWith(expect.stringContaining("Failed to update user ["));
         });
     });
